backlog 3/6: Int → Long primary keys (schema v3)
Deck.id / Flashcard.id / Flashcard.deckId / ChatLog.id / ChatLog.deckId /
ChatLog.flashcardId are Long. SQLite stores every INTEGER column as a 64-bit
value already, so Room's generated schema is byte-identical (v2 and v3 share
identity hash fcf2ec48…) — MIGRATION_2_3 is a no-op that only exists to let
Room accept the version bump. Zero data-migration risk.

Ripples: DAO/repo params, _flippedCardIds Set<Long>, _cardModes Map<Long,_>,
getLearningModeForCard(Long), savedState KEY_ACTIVE_DECK as Long, ImportUndo
id lists Long, dropped now-redundant .toInt() on insert results, DeckMergeTest
card ids .toLong().

Device-verified: install v3 over v2 → app launches (1.7s), Master Pool's 142
cards intact, study loop grades + writes fine, no Room/SQLite/cast errors.

+++ app/src/main/java/io/github/playfoundryhq/adaptiveflow/data/dao/StudyDao.kt
@@ -22,7 +22,7 @@ interface StudyDao {
     fun getAllDecksWithCardsFlow(): Flow<List<io.github.playfoundryhq.adaptiveflow.data.model.DeckWithCards>>
 
     @Query("SELECT * FROM decks WHERE id = :id")
-    suspend fun getDeckById(id: Int): Deck?
+    suspend fun getDeckById(id: Long): Deck?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertDeck(deck: Deck): Long
@@ -35,13 +35,13 @@ interface StudyDao {
 
     // --- Flashcards ---
     @Query("SELECT * FROM flashcards WHERE deckId = :deckId")
-    fun getFlashcardsForDeckFlow(deckId: Int): Flow<List<Flashcard>>
+    fun getFlashcardsForDeckFlow(deckId: Long): Flow<List<Flashcard>>
 
     @Query("SELECT * FROM flashcards WHERE deckId = :deckId")
-    suspend fun getFlashcardsForDeck(deckId: Int): List<Flashcard>
+    suspend fun getFlashcardsForDeck(deckId: Long): List<Flashcard>
 
     @Query("SELECT * FROM flashcards WHERE id = :id")
-    suspend fun getFlashcardById(id: Int): Flashcard?
+    suspend fun getFlashcardById(id: Long): Flashcard?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertFlashcard(flashcard: Flashcard): Long
@@ -57,16 +57,16 @@ interface StudyDao {
 
     // --- Chat Logs ---
     @Query("SELECT * FROM chat_logs WHERE deckId = :deckId ORDER BY timestamp ASC")
-    fun getChatLogsForDeckFlow(deckId: Int): Flow<List<ChatLog>>
+    fun getChatLogsForDeckFlow(deckId: Long): Flow<List<ChatLog>>
 
     @Query("SELECT * FROM chat_logs WHERE deckId = :deckId ORDER BY timestamp ASC")
-    suspend fun getChatLogsForDeck(deckId: Int): List<ChatLog>
+    suspend fun getChatLogsForDeck(deckId: Long): List<ChatLog>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertChatLog(chatLog: ChatLog): Long
 
     @Query("DELETE FROM chat_logs WHERE deckId = :deckId")
-    suspend fun clearChatLogsForDeck(deckId: Int)
+    suspend fun clearChatLogsForDeck(deckId: Long)
 
     // --- Progress (single row, id = 0) ---
     @Query("SELECT * FROM progress WHERE id = 0")
